Adapted files for C55 compiling

+++ Decoder/inc/fft.h
@@ -9,6 +9,6 @@
 #define FFT_H_
 
 void fft(double* samples, double* realCoeficient, double* imCoeficient, int totalsamples);
-void ifft(double* realCoeficient, double* imCoeficient, double* SamplesIFFT, int totalsamples);
+void ifft(int* realCoeficient, int* imCoeficient, double* SamplesIFFT, int totalsamples);
 
 #endif
